seeder fixed.

+++ database/migrations/2022_11_14_160945_create_teacher_infos_table.php
@@ -17,10 +17,10 @@ class CreateTeacherInfosTable extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id')->constrained()->onDelete('cascade')->index();
             $table->string('username')->unique()->nullable()->index();
-            $table->integer('bank_account_name')->nullable();
-            $table->integer('bank_account_no')->nullable();
-            $table->integer('bank_account_branch')->nullable();
-            $table->integer('nid')->nullable();
+            $table->bigInteger('bank_account_name')->nullable();
+            $table->bigInteger('bank_account_no')->nullable();
+            $table->bigInteger('bank_account_branch')->nullable();
+            $table->bigInteger('nid')->nullable();
             $table->string('nid_img')->nullable();
             $table->string('institute')->nullable();
             $table->string('curriculum')->nullable();
